requirements.txt
tensorflow
numpy
Pillow
sklearn
tensorflow_datasets
plac